loading data

+++ Weather_App/WeatherClass.swift
@@ -37,8 +37,8 @@ class WeatherClass: NSObject {
     
     
     //Coordinate
-    var latitude: Float = 0.0
-    var longitude: Float = 0.0
+    var latitude: Double = 0.0
+    var longitude: Double = 0.0
     
     
     //weather
